requirements.txt
streamlit == 1.30.0
pandas == 2.1.4
seaborn == 0.12.2
matplotlib == 3.7.2
numpy== 1.24.3
scikit-learn == 1.3.0
plotly == 5.9.0
Pillow == 10.0.1
feature_engine == 1.6.2
streamlit-option-menu==0.3.12
roboflow == 1.1.31